--- a/build.gradle
+++ b/build.gradle
@@ -134,6 +134,7 @@ def integrationTest = tasks.register("integrationTest", Test) {
                 options.includeTags 'integration'
         }
         task.shouldRunAfter test
+        task.maxParallelForks(Runtime.runtime.availableProcessors().intdiv(2) ?: 1)
 }
 
 def testAll = tasks.register("testAll", Test) {