This is a branch for various experiments I'm trying to speed up the tests

## Improvements
- Make `UserJourneyPageTest` and `PdfIntegrationTest` unit tests to enable greater parallelization
  - One of the `UserJourneyPageTest`s could not be parallelized so I extracted it into `NonParallelizableUserJourneyPageTest`, which is still an integration test
- Removed tests from `UserJourneyPageTest` that are fully covered by other tests:
  - removed `userCanCompleteTheNonExpeditedFlow` and `shouldDownloadPDFWhenClickDownloadMyReceipt` as they could both be merged into `userCanCompleteTheNonExpeditedFlowAndCanDownloadPdfsAndShibaShouldCaptureMetricsAfterApplicationIsCompleted`. (Nasty name but I think it's reasonable give the potential speed up)
- Moved many helpers from `UserJourneyPageTest` into `AbstractBasePageTest`, making them available to other test files
- Split up `UserJourneyPageTest` into multiple tests that can be run in parallel (thanks to @bencalegari for doing this!)

## Fixes
- Made `SubmitPageTest` an integration test, as it causes test pollution by truncating the `applications` table

## Benchmarkes
All measured on my machine using

```
./gradlew clean testAll --scan --rerun-tasks
```

* Baseline: 15m 14s (https://scans.gradle.com/s/homrplsyovurg)
* ~After parallelizing Integration tests: 15m 5s (https://scans.gradle.com/s/6xn6fhji2nif4)~
* Making `UserJourneyPageTest` and `PdfIntegrationTest` into unit tests: 11m 22s (https://scans.gradle.com/s/qk4oudxingske)
* Making `NonParallelizableUserJourneyPageTest` and `SubmitPageTest` into integration tests to avoid test flakes: 12m 16s (https://scans.gradle.com/s/66mb4nmhx22qy)
  * The slowdown is a bummer but this makes the rest of the changes safe.
* Split up `UserJourneyPageTest`: **10m 5s!!!** (https://scans.gradle.com/s/lygo7mwf4w2dc)